requirements.txt
fastapi>=0.100.0
fastapi[standard]>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
loguru>=0.7.0
astropy>=5.3.0
numpy>=1.24.0
pandas>=2.3.3
alpyca>=3.1.1